+++ src/appnet-preview/azext_appnet_preview/aaz/latest/appnet/_list.py
@@ -29,10 +29,10 @@ class List(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2025-08-01-preview",
+        "version": "2026-08-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.applink/applinks", "2025-08-01-preview"],
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.applink/applinks", "2025-08-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.applink/applinks", "2026-08-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.applink/applinks", "2026-08-01-preview"],
         ]
     }
 
@@ -119,7 +119,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-08-01-preview",
+                    "api-version", "2026-08-01-preview",
                     required=True,
                 ),
             }
@@ -289,7 +289,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-08-01-preview",
+                    "api-version", "2026-08-01-preview",
                     required=True,
                 ),
             }