+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/accessbridge/_create.py
@@ -22,9 +22,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2026-07-01",
+        "version": "2026-08-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/accessbridges/{}", "2026-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/accessbridges/{}", "2026-08-01-preview"],
         ]
     }
 
@@ -49,7 +49,7 @@ class Create(AAZCommand):
             options=["-n", "--name", "--access-bridge-name"],
             help="The name of the access bridge.",
             required=True,
-            enum={"Bastion": "Bastion", "PrivateVault": "PrivateVault", "StorageDashboard": "StorageDashboard"},
+            enum={"Bastion": "Bastion", "EdgeManagement": "EdgeManagement", "PrivateVault": "PrivateVault", "StorageDashboard": "StorageDashboard"},
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -235,7 +235,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2026-07-01",
+                    "api-version", "2026-08-01-preview",
                     required=True,
                 ),
             }