+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/storageappliance/_show.py
@@ -23,9 +23,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2026-07-01",
+        "version": "2026-08-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/storageappliances/{}", "2026-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/storageappliances/{}", "2026-08-01-preview"],
         ]
     }
 
@@ -124,7 +124,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2026-07-01",
+                    "api-version", "2026-08-01-preview",
                     required=True,
                 ),
             }
@@ -332,6 +332,10 @@ class Show(AAZCommand):
             )
 
             secret_archive_reference = cls._schema_on_200.properties.secret_rotation_status.Element.secret_archive_reference
+            secret_archive_reference.encryption_public_key = AAZStrType(
+                serialized_name="encryptionPublicKey",
+                flags={"read_only": True},
+            )
             secret_archive_reference.key_vault_id = AAZStrType(
                 serialized_name="keyVaultId",
                 flags={"read_only": True},